--- conifer-config.json
@@ -1,1 +1,0 @@
-{"awsRegion":"us-west-1","testDirectory":"/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress","packageManager":"yarn","entryPoint":"yarn dev","ports":["3000","3001"],"parallelInstances":"4","ec2InstanceType":"c6i.4xlarge","parallelType":"File globbing","bucketName":"conifer-test-bucket-17533d18-a100-46da-91b1-cd7d7cd039d9","testGroupings":[["/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-bankaccounts.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-likes.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-users.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui-auth-providers/google.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/new-transaction.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/user-settings.spec.ts"],["/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-banktransfers.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-notifications.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/demo/cypress-studio.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui-auth-providers/okta.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/notifications.spec.ts"],["/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-comments.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-testdata.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui-auth-providers/auth0.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/auth.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/transaction-feeds.spec.ts"],["/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-contacts.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/api/api-transactions.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui-auth-providers/cognito.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/bankaccounts.spec.ts","/Users/lawrence/Documents/Capstone/cypress-test/cypress-realworld-app/cypress/tests/ui/transaction-view.spec.ts"]],"imageUri":"988964088119.dkr.ecr.us-west-1.amazonaws.com/conifer-test:latest","testRunId":"def8b1ac-2563-4cc9-a83d-99119dfd9361","taskRunARNs":["arn:aws:ecs:us-west-1:988964088119:task/ConiferCdkStack-coniferECSclusterC14FBDF3-szoW01sFuRA5/7267dad32c824e088cb0413fec5dce15","arn:aws:ecs:us-west-1:988964088119:task/ConiferCdkStack-coniferECSclusterC14FBDF3-szoW01sFuRA5/c536c590990a46388fc6931b8ec283ba","arn:aws:ecs:us-west-1:988964088119:task/ConiferCdkStack-coniferECSclusterC14FBDF3-szoW01sFuRA5/cd676067a2554f0ebf034b8576b09c43","arn:aws:ecs:us-west-1:988964088119:task/ConiferCdkStack-coniferECSclusterC14FBDF3-szoW01sFuRA5/4d06b3ff7be24037bff46c68ebab0bea"]}
