--- server/.env
@@ -1,1 +1,0 @@
-AWS_REGION=ap-northeast-1
