--- server/controllers/webHookController.js
@@ -1,18 +1,0 @@
-// const { createSession } = require('better-sse');
-
-// let sseSession;
-// // app.post('/api/webhookURL', (request, response) => {
-// //   const { id, status } = request.body;
-
-// //   sseSession.push(phoneCallsStatus[id]);
-
-// //   response.status(200).end();
-// // });
-
-// const updateTestFile = (req, res, next) => {
-//   const { data } = req.body;
-
-//   sseSession.push(data);
-
-//   response.status(200).end();
-// };
