+++ server/index.js
@@ -1,11 +1,10 @@
 const express = require('express');
+const { createSession } = require('better-sse');
+const fs = require('fs');
 const routes = require('./routes/api');
 const HttpError = require('./models/httpError');
-require('dotenv').config();
-const { areTasksRunning } = require('./utils/areTasksRunning');
 const pollDynamoDb = require('./utils/pollDynamoForUpdates');
-const fs = require('fs');
-const { createSession } = require('better-sse');
+require('dotenv').config();
 
 const app = express();
 let sseSession;
@@ -54,20 +53,10 @@ app.use((err, req, res, next) => {
 
 const port = process.env.PORT || 5001;
 
-// For testing purposes
 app.listen(port, async () => {
-  const { testRunId, taskRunARNs } = JSON.parse(
-    fs.readFileSync('../../conifer-config.json')
-  );
-
-  console.log(`Server running on port ${port}`); //ORIGINAL
-  // console.log('taskArns: ', taskArns);
-
-  const areTasksTest = await areTasksRunning(taskRunARNs);
-  console.log('areTasksTest: ', areTasksTest);
-
-  // const res = await getItemsByTestRunID('c2a72ecf-ad30-44b0-a035-130e527b8457');
-  // console.log(res);
-  // pollDynamoDb('c2a72ecf-ad30-44b0-a035-130e527b8457', taskArns);
-  pollDynamoDb(testRunId, taskRunARNs);
+  // const { testRunId, taskRunARNs } = JSON.parse(
+  //   fs.readFileSync('../../conifer-config.json')
+  // );
+  console.log(`Server running on port ${port}`);
+  // pollDynamoDb(testRunId, taskRunARNs);
 });
